requirements.txt
streamlit
google-generativeai>=0.8.6
google-api-core>=2.17.1
google-auth>=2.27.0
firebase-admin>=6.5.0
requests
pandas
openpyxl
pypdf
watchdog
# tavily-python  <-- Descomentar solo si vas a configurar la API de Tavily para búsquedas